refactor: extract search product bar
