+++ src/pages/Retail.jsx
@@ -1,11 +1,11 @@
 /* eslint-disable no-unused-vars */
 import { Button, Form } from "antd";
 import { useDispatch } from "react-redux";
-import SearchProductBar from "../features/products/SearchProductBar";
 import OrderDetailsTable from "../features/orders/OrderDetailsTable";
 import SearchCustomerBar from "../features/customers/SearchCustomerBar";
 import CreateOrderForm from "../features/orders/CreateOrderForm";
 import { clearOrderDetails } from "../features/orders/orderSlice";
+import SearchProduct from "../features/products/SearchProduct";
 
 function Retail() {
   const dispatch = useDispatch();
@@ -14,7 +14,7 @@ function Retail() {
     <div className="flex items-center justify-between">
       <div className="card min-h-[calc(100vh-64px-1.5rem*2)] basis-[68%] space-y-8">
         <div className="flex items-center justify-between">
-          <SearchProductBar />
+          <SearchProduct />
           <Button
             type="primary"
             danger
